Refactor grid style types and impls
Fixes #16949.
