+++ components/style/values/generics/mod.rs
@@ -16,6 +16,7 @@ use super::CustomIdent;
 pub use self::basic_shape::serialize_radius_values;
 
 pub mod basic_shape;
+pub mod grid;
 pub mod image;
 pub mod position;
